Move localStorageIsAvailable util to its own file
utils/index.ts has a dependency on legacy web3js
